Refactor web-specific code to use conditional imports and stubs
