+++ church_on_app/lib/common/services/domain_service_stub.dart
@@ -1,0 +1,3 @@
+class DomainService {
+  Future<String?> resolveChurchIdFromHost() async => null;
+}
