+++ church_on_app/lib/common/services/domain_service_web.dart
@@ -1,0 +1,17 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+// ignore: avoid_web_libraries_in_flutter
+import 'dart:html' as html;
+
+class DomainService {
+  DomainService({FirebaseFirestore? firestore}) : _firestore = firestore ?? FirebaseFirestore.instance;
+  final FirebaseFirestore _firestore;
+
+  Future<String?> resolveChurchIdFromHost() async {
+    final host = html.window.location.host.toLowerCase();
+    if (host.isEmpty) return null;
+    final snap = await _firestore.collection('domains').doc(host).get();
+    if (!snap.exists) return null;
+    final data = snap.data() as Map<String, dynamic>;
+    return data['churchId'] as String?;
+  }
+}
