+++ church_on_app/lib/common/web/export_csv.dart
@@ -1,0 +1,3 @@
+void exportCsv(String filename, List<int> bytes) {
+  // no-op on non-web; tests can call this safely
+}
