+++ church_on_app/lib/common/web/export_csv_web.dart
@@ -1,0 +1,14 @@
+// ignore: avoid_web_libraries_in_flutter
+import 'dart:html' as html;
+
+void exportCsv(String filename, List<int> bytes) {
+  final blob = html.Blob([bytes], 'text/csv');
+  final url = html.Url.createObjectUrlFromBlob(blob);
+  final anchor = html.document.createElement('a') as html.AnchorElement
+    ..href = url
+    ..download = filename;
+  html.document.body!.append(anchor);
+  anchor.click();
+  anchor.remove();
+  html.Url.revokeObjectUrl(url);
+}
